+++ tests/testthat/test_intersecting_units.R
@@ -3,9 +3,12 @@ context("intersecting_units")
 test_that("x=Raster,y=Raster", {
   # generate data
   x <- raster::raster(matrix(c(NA, 2:16), byrow = TRUE, ncol = 4))
+  x2 <- raster::stack(x)
   y <- x > 6
+  y2 <- raster::stack(y)
   # run tests
   expect_equal(intersecting_units(x, y), c(7:16))
+  expect_equal(intersecting_units(x, y), intersecting_units(x2, y2))
   # check that invalid arguments result in errors
   y_crs <- `crs<-`(y, value = sp::CRS("+proj=longlat +datum=WGS84 +no_defs"))
   expect_error(intersecting_units(x, y_crs))
@@ -88,8 +91,10 @@ test_that("x=sf,y=Raster", {
   m <- mean(sim_pu_sf$cost, na.rm = TRUE)
   x <- sim_pu_sf
   y <- sim_pu_raster < m
+  y2 <- raster::stack(y)
   # run tests
   expect_equal(intersecting_units(x, y), which(sim_pu_sf$cost < m))
+  expect_equal(intersecting_units(x, y), intersecting_units(x, y2))
   # check that invalid arguments result in errors
   y_crs <- `crs<-`(y, value = sp::CRS("+proj=longlat +datum=WGS84 +no_defs"))
   expect_error(intersecting_units(x, y_crs))
@@ -119,4 +124,8 @@ test_that("x=Spatial,y=sf", {
   y_crs <- sf::st_set_crs(y, 4326)
   expect_error(intersecting_units(x, y_crs))
   expect_error(intersecting_units(x[1:5, ], x[50:55, ]))
+})
+
+test_that("invalid input", {
+  expect_error(intersecting_units(sim_pu_polygons@data, sim_pu_polygons))
 })