+++ NEWS.md
@@ -8,6 +8,8 @@
 - Update documentation with information about Ferrier importance scores.
 - Brandon Edwards added to package author list.
 - Update Gurobi Installation guide vignette.
+- Update `presolve_check` function to throw warning when
+  really high values specified in `add_neighbor_constraints`.
 
 # prioritizr 7.1.0.3